refactor: separate publish from pm subcommands
