+++ cli/registry.rs
