+++ cli/tools/installer.rs
@@ -212,7 +212,7 @@ pub async fn uninstall(
   let uninstall_flags = match uninstall_flags.kind {
     UninstallKind::Global(flags) => flags,
     UninstallKind::Local(remove_flags) => {
-      return super::registry::remove(flags, remove_flags).await;
+      return super::pm::remove(flags, remove_flags).await;
     }
   };
 
@@ -300,12 +300,7 @@ async fn install_local(
 ) -> Result<(), AnyError> {
   match install_flags {
     InstallFlagsLocal::Add(add_flags) => {
-      super::registry::add(
-        flags,
-        add_flags,
-        super::registry::AddCommandName::Install,
-      )
-      .await
+      super::pm::add(flags, add_flags, super::pm::AddCommandName::Install).await
     }
     InstallFlagsLocal::Entrypoints(entrypoints) => {
       install_from_entrypoints(flags, &entrypoints).await
@@ -314,7 +309,7 @@ async fn install_local(
       let factory = CliFactory::from_flags(flags);
       // surface any errors in the package.json
       factory.npm_installer()?.ensure_no_pkg_json_dep_errors()?;
-      crate::tools::registry::cache_top_level_deps(&factory, None).await?;
+      crate::tools::pm::cache_top_level_deps(&factory, None).await?;
 
       if let Some(lockfile) = factory.cli_options()?.maybe_lockfile() {
         lockfile.write_if_changed()?;
@@ -393,8 +388,7 @@ async fn install_global(
   let entry_text = install_flags_global.module_url.as_str();
   if !cli_options.initial_cwd().join(entry_text).exists() {
     // check for package requirement missing prefix
-    if let Ok(Err(package_req)) =
-      super::registry::AddRmPackageReq::parse(entry_text)
+    if let Ok(Err(package_req)) = super::pm::AddRmPackageReq::parse(entry_text)
     {
       if jsr_resolver.req_to_nv(&package_req).await.is_some() {
         bail!(
