+++ cli/tools/mod.rs
@@ -12,7 +12,8 @@ pub mod init;
 pub mod installer;
 pub mod jupyter;
 pub mod lint;
-pub mod registry;
+pub mod pm;
+pub mod publish;
 pub mod repl;
 pub mod run;
 pub mod serve;
