+++ cli/tools/pm/outdated/mod.rs
@@ -16,6 +16,7 @@ use deno_terminal::colors;
 
 use super::deps::Dep;
 use super::deps::DepId;
+use super::deps::DepKind;
 use super::deps::DepManager;
 use super::deps::DepManagerArgs;
 use super::deps::PackageLatestVersion;
@@ -26,7 +27,6 @@ use crate::factory::CliFactory;
 use crate::file_fetcher::CliFileFetcher;
 use crate::jsr::JsrFetchResolver;
 use crate::npm::NpmFetchResolver;
-use crate::tools::registry::pm::deps::DepKind;
 
 #[derive(Debug, PartialEq, Eq, PartialOrd, Ord)]
 struct OutdatedPackage {
