+++ cli/tools/publish/auth.rs
