+++ cli/tools/publish/module_content.rs
@@ -16,6 +16,7 @@ use lazy_regex::Lazy;
 use sys_traits::FsMetadata;
 use sys_traits::FsRead;
 
+use super::diagnostics::PublishDiagnostic;
 use super::diagnostics::PublishDiagnosticsCollector;
 use super::unfurl::SpecifierUnfurler;
 use super::unfurl::SpecifierUnfurlerDiagnostic;
@@ -23,7 +24,6 @@ use crate::args::deno_json::TsConfigResolver;
 use crate::cache::LazyGraphSourceParser;
 use crate::cache::ParsedSourceCache;
 use crate::sys::CliSys;
-use crate::tools::registry::diagnostics::PublishDiagnostic;
 
 struct JsxFolderOptions<'a> {
   jsx_factory: &'a str,
