+++ cli/tools/publish/paths.rs
